add one-link cmd
