+++ lib/url.js
@@ -17,15 +17,7 @@ function splitFromBase64(token) {
   return [password, passcode, network];
 }
 
-async function main() {
-  const token = mergeToBase64("DemoAccount5&", "112324", "polygon");
-  console.log("token: ", token);
-
-  const [password, passcode, network] = splitFromBase64(token);
-  console.log(password, passcode, network);
-}
-
-main().then(() => {
-  console.log("\nfinished.");
-  process.exit(0);
-});
+module.exports = {
+  mergeToBase64,
+  splitFromBase64,
+};
