+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "balances": "node index.js",
-    "link": "node link.js"
+    "one-link": "node one-link.js"
   },
   "author": "",
   "license": "ISC",
